requirements.txt
PyQt6>=6.5.0
opencv-python>=4.8.0
easyocr>=1.7.0
Pillow>=10.0.0
webptools
SQLAlchemy>=2.0.0
asyncio>=3.4.3
PyQt6-WebEngine>=6.5.0
numpy>=1.24.0